oms2railjson: move adjacency code to its own file
